feat!(sdk): renames opentdf/client opentdf/sdk (#374)
- Renames library from `client` to `sdk` to reflect we are adding more platform-specific features, in addition to the core TDF functionality
- Removes remote-store, as it is unmaintained
- Resets version numbering to 0.1.0
- tags from here on out will be prefixed with library short name and a slash
- renames cli to ctl so we can keep lockstep versioning the different packages
