+++ .github/workflows/format.yaml
@@ -20,10 +20,10 @@ jobs:
       - run: git diff-files --ignore-submodules
       - name: Check that files have been formatted before PR submission
         run: git diff-files --quiet --ignore-submodules
-        if: ${{ github.event.pull_request.head.repo.full_name != 'opentdf/client-web' }}
+        if: ${{ github.event.pull_request.head.repo.full_name != 'opentdf/web-sdk' }}
       - name: Commit changes
         id: auto-commit
-        if: ${{ github.event.pull_request.head.repo.full_name == 'opentdf/client-web' }}
+        if: ${{ github.event.pull_request.head.repo.full_name == 'opentdf/web-sdk' }}
         uses: stefanzweifel/git-auto-commit-action@v5
         with:
           commit_message: |-
