+++ .github/workflows/gh-semver.sh
@@ -29,7 +29,7 @@
 # Tags go to release:
 # ```
 #    package.version = 1.2.3
-#    tag = v1.2.3
+#    tag = sdk/v1.2.3
 #    workflow run = 256
 #    git SHA = decaf
 #    ----
