+++ .github/workflows/guess-dist-tag.bats
@@ -29,7 +29,7 @@
 }
 
 @test "all tags go to release" {
-  export GITHUB_REF=refs/tags/v12
+  export GITHUB_REF=refs/tags/sdk/v12
   run $BATS_TEST_DIRNAME/guess-dist-tag.sh
   echo output=[$output]
   [[ $output == "latest" ]]
