+++ .github/workflows/guess-dist-tag.sh
@@ -1,8 +1,12 @@
 #!/usr/bin/env bash
-# Guess the desired NPM 'dist' tag based on current git ref
-# Release = latest
-# Release candidate = rc
-# Beta = beta
+# Guess the desired NPM distribution tag based on current git ref.
+# For more info, see: https://docs.npmjs.com/adding-dist-tags-to-packages
+# Releases are tagged with `latest`, on tags like `sdk/v1.2.3`
+# Release candidates are tagged `rc`, from branches prefixed with `release/`
+# Betas are the main branch.
+# Alphas can be manually built from feature branches.
+# Aleph is the fallback for unknown branch and tag patterns.
+# Notably, our dist-tags sort lexicographically from least to most stable.
 
 set -euo pipefail
 
@@ -19,7 +23,7 @@ case "${GITHUB_REF}" in
   refs/heads/feature*)
     NPM_DIST_TAG=alpha
     ;;
-  refs/tags/v*)
+  refs/tags/sdk/v*)
     NPM_DIST_TAG=latest
     ;;
 esac
