+++ .github/workflows/roundtrip/encrypt-decrypt.sh
@@ -35,7 +35,7 @@ _nano_test() {
   rm -f "${plain}" sample.txt.ntdf sample_out.txt
 }
 
-_nano_test @opentdf/cli @opentdf/cli
+_nano_test @opentdf/ctl @opentdf/ctl
 
 _tdf3_test() {
   counter=$((counter + 1))
@@ -66,4 +66,4 @@ _tdf3_test() {
   rm -f "${plain}" sample.txt.tdf sample_out.txt
 }
 
-_tdf3_test @opentdf/cli @opentdf/cli
+_tdf3_test @opentdf/ctl @opentdf/ctl
