+++ .github/workflows/roundtrip/package.json
@@ -1,9 +1,9 @@
 {
-    "name": "client-web-roundtrip",
+    "name": "web-sdk-roundtrip",
     "version": "0.0.1",
     "description": "Simple example to encrypt and decrypt files with quickstart backend.",
     "scripts": {},
     "dependencies": {
-        "@opentdf/cli": "file:../../../cli/opentdf-cli-2.0.0.tgz"
+        "@opentdf/ctl": "file:../../../cli/opentdf-ctl-0.1.0.tgz"
     }
 }
