+++ cli/bin/opentdf.bats
@@ -22,6 +22,6 @@
 @test "version command" {
   run $BATS_TEST_DIRNAME/opentdf.mjs --version
   echo "$output"
-  [[ $output == *"@opentdf/client\":\""* ]]
-  [[ $output == *"@opentdf/cli\":\""* ]]
+  [[ $output == *"@opentdf/sdk\":\""* ]]
+  [[ $output == *"@opentdf/ctl\":\""* ]]
 }
