+++ lib/src/version.ts
@@ -1,9 +1,9 @@
 /**
- * Exposes the released version number of the `@opentdf/client` package
+ * Exposes the released version number of the `@opentdf/sdk` package
  */
-export const version = '2.1.0';
+export const version = '0.1.0';
 
 /**
  * A string name used to label requests as coming from this library client.
  */
-export const clientType = 'client-web';
+export const clientType = 'web-sdk';
