+++ lib/tdf3/src/version.ts
@@ -1,2 +1,2 @@
-export const version = '2.1.0';
+export const version = '0.1.0';
 export const clientType = 'tdf3-js-client';
