--- remote-store/.eslintrc.yaml
@@ -1,13 +1,0 @@
-env:
-  browser: true
-extends:
-  - plugin:@typescript-eslint/recommended
-  - prettier
-overrides:
-  - files:
-      - '*.spec.ts'
-    rules:
-      '@typescript-eslint/ban-ts-comment': 'off'
-parserOptions:
-  project: ['**/tsconfig.json']
-root: true
