--- remote-store/README.md
@@ -1,3 +1,0 @@
-# @opentdf/remote-store
-
-This library extends the @opentdf/client library with functions to enable streaming to and from S3 buckets and other storage services that implement the S3 API.
