--- remote-store/src/aws-lib-storage/chunks/getDataReadable.ts
@@ -1,5 +1,0 @@
-export async function* getDataReadable(data: any): AsyncGenerator<Uint8Array> {
-  for await (const chunk of data) {
-    yield chunk;
-  }
-}
