--- remote-store/src/aws-lib-storage/chunks/getDataReadableStream.ts
@@ -1,18 +1,0 @@
-export async function* getDataReadableStream(data: ReadableStream): AsyncGenerator<Uint8Array> {
-  // Get a lock on the stream.
-  const reader = data.getReader();
-
-  try {
-    while (true) {
-      // Read from the stream.
-      const { done, value } = await reader.read();
-      // Exit if we're done.
-      if (done) return;
-      // Else yield the chunk.
-      yield value;
-    }
-  } finally {
-    // release the lock for reading from this stream.
-    reader.releaseLock();
-  }
-}
