--- remote-store/src/aws-lib-storage/index.ts
@@ -1,4 +1,0 @@
-export { Upload } from './Upload.js';
-// its a aws lib https://github.com/aws/aws-sdk-js-v3/tree/main/lib/lib-storage
-// had to move it here to make it work with
-// https://developer.salesforce.com/docs/component-library/tools/locker-service-console
