--- remote-store/src/aws-lib-storage/runtimeConfig.browser.ts
@@ -1,9 +1,0 @@
-import { ClientSharedValues } from './runtimeConfig.shared.js';
-
-/**
- * @internal
- */
-export const ClientDefaultValues = {
-  ...ClientSharedValues,
-  runtime: 'browser',
-};
