--- remote-store/src/aws-lib-storage/runtimeConfig.shared.ts
@@ -1,7 +1,0 @@
-/**
- * @internal
- */
-export const ClientSharedValues = {
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
-  lstatSync: () => {},
-};
