--- remote-store/src/aws-lib-storage/runtimeConfig.ts
@@ -1,12 +1,0 @@
-import { lstatSync } from 'node:fs';
-
-import { ClientSharedValues } from './runtimeConfig.shared.js';
-
-/**
- * @internal
- */
-export const ClientDefaultValues = {
-  ...ClientSharedValues,
-  runtime: 'node',
-  lstatSync,
-};
