--- remote-store/tsconfig.json
@@ -1,32 +1,0 @@
-{
-  "compilerOptions": {
-    "alwaysStrict": true,
-    "baseUrl": "./",
-    "composite": true,
-    "declaration": true,
-    "declarationDir": "./dist",
-    "declarationMap": true,
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "inlineSourceMap": true,
-    "isolatedModules": true,
-    "lib": ["dom", "dom.iterable", "es2020", "ES2022.Error"],
-    "module": "es2020",
-    "moduleResolution": "node16",
-    "noImplicitAny": true,
-    "noImplicitOverride": true,
-    "noImplicitReturns": true,
-    "noImplicitThis": true,
-    "noUnusedLocals": true,
-    "outDir": "dist",
-    "skipLibCheck": true,
-    "strict": false,
-    "strictBindCallApply": true,
-    "strictFunctionTypes": true,
-    "strictNullChecks": true,
-    "strictPropertyInitialization": false,
-    "target": "es2020",
-  },
-  "include": ["src", "tests"],
-  "exclude": ["**/node_modules"]
-}
