+++ web-app/package.json
@@ -1,6 +1,6 @@
 {
   "name": "web-app",
-  "version": "2.1.0",
+  "version": "0.1.0",
   "license": "BSD-3-Clause-Clear",
   "type": "module",
   "scripts": {
@@ -15,7 +15,7 @@
     "test:ui": "vite build && vitest --ui"
   },
   "dependencies": {
-    "@opentdf/client": "file:../lib/opentdf-client-2.1.0.tgz",
+    "@opentdf/sdk": "file:../lib/opentdf-sdk-0.1.0.tgz",
     "clsx": "^2.0.0",
     "native-file-system-adapter": "^3.0.1",
     "react": "^18.2.0",
